+++ src/main/java/ru/vyarus/guice/persist/orient/support/AutoScanFinderModule.java
@@ -16,7 +16,11 @@ import java.util.List;
  *
  * @author Vyacheslav Rusakov
  * @since 04.08.2014
+ * @deprecated finders now completely
+ * <a href="https://github.com/xvik/guice-ext-annotations#usage">controlled by guice</a>
+ * you can use @ProvidedBy instead of classpath scanning
  */
+@Deprecated
 public class AutoScanFinderModule extends FinderModule {
     private final Logger logger = LoggerFactory.getLogger(AutoScanFinderModule.class);
     private final String[] packages;