+++ src/test/groovy/ru/vyarus/guice/persist/orient/finder/ExtraCasesExecutionTest.groovy
@@ -5,14 +5,14 @@ import ru.vyarus.guice.persist.orient.AbstractTest
 import ru.vyarus.guice.persist.orient.db.transaction.template.SpecificTxAction
 import ru.vyarus.guice.persist.orient.support.finder.ExtraCasesFinder
 import ru.vyarus.guice.persist.orient.support.model.Model
-import ru.vyarus.guice.persist.orient.support.modules.AutoScanFinderTestModule
+import ru.vyarus.guice.persist.orient.support.modules.FinderTestModule
 import spock.guice.UseModules
 
 /**
  * @author Vyacheslav Rusakov 
  * @since 05.08.2014
  */
-@UseModules(AutoScanFinderTestModule)
+@UseModules(FinderTestModule)
 class ExtraCasesExecutionTest extends AbstractTest {
 
     @Inject