+++ src/test/groovy/ru/vyarus/guice/persist/orient/finder/FinderCrudMixinsTest.groovy
@@ -12,13 +12,14 @@ import ru.vyarus.guice.persist.orient.support.finder.ObjectDao
 import ru.vyarus.guice.persist.orient.support.finder.mixin.pagination.Page
 import ru.vyarus.guice.persist.orient.support.model.Model
 import ru.vyarus.guice.persist.orient.support.modules.AutoScanFinderTestModule
+import ru.vyarus.guice.persist.orient.support.modules.FinderTestModule
 import spock.guice.UseModules
 
 /**
  * @author Vyacheslav Rusakov 
  * @since 26.10.2014
  */
-@UseModules(AutoScanFinderTestModule)
+@UseModules(FinderTestModule)
 class FinderCrudMixinsTest extends AbstractTest {
 
     @Inject