+++ src/test/groovy/ru/vyarus/guice/persist/orient/finder/internal/ExtraCasesAnalysisTest.groovy
@@ -8,13 +8,14 @@ import ru.vyarus.guice.persist.orient.finder.result.ResultType
 import ru.vyarus.guice.persist.orient.support.finder.ExtraCasesFinder
 import ru.vyarus.guice.persist.orient.support.model.Model
 import ru.vyarus.guice.persist.orient.support.modules.AutoScanFinderTestModule
+import ru.vyarus.guice.persist.orient.support.modules.FinderTestModule
 import spock.guice.UseModules
 
 /**
  * @author Vyacheslav Rusakov 
  * @since 05.08.2014
  */
-@UseModules(AutoScanFinderTestModule)
+@UseModules(FinderTestModule)
 class ExtraCasesAnalysisTest extends AbstractFinderDefinitionTest {
 
     def "Check db type recognition"() {