+++ src/test/groovy/ru/vyarus/guice/persist/orient/support/finder/DocumentDao.groovy
@@ -1,5 +1,7 @@
 package ru.vyarus.guice.persist.orient.support.finder
 
+import com.google.inject.ProvidedBy
+import com.google.inject.internal.DynamicClassProvider
 import com.google.inject.persist.Transactional
 import com.orientechnologies.orient.core.record.impl.ODocument
 import ru.vyarus.guice.persist.orient.support.finder.mixin.crud.DocumentCrudMixin
@@ -11,6 +13,7 @@ import ru.vyarus.guice.persist.orient.support.model.Model
  * @since 26.10.2014
  */
 @Transactional
+@ProvidedBy(DynamicClassProvider)
 public interface DocumentDao extends DocumentCrudMixin<Model>,
         CustomMixin<ODocument, String>,
         PaginationMixin<Model, ODocument> {