+++ src/test/groovy/ru/vyarus/guice/persist/orient/support/finder/ExtraCasesFinder.groovy
@@ -1,5 +1,7 @@
 package ru.vyarus.guice.persist.orient.support.finder
 
+import com.google.inject.ProvidedBy
+import com.google.inject.internal.DynamicClassProvider
 import com.google.inject.persist.Transactional
 import com.google.inject.persist.finder.Finder
 import com.orientechnologies.orient.core.record.impl.ODocument
@@ -13,6 +15,7 @@ import ru.vyarus.guice.persist.orient.support.model.Model
  * @since 05.08.2014
  */
 @Transactional
+@ProvidedBy(DynamicClassProvider)
 interface ExtraCasesFinder {
 
     // iterable result