+++ src/test/groovy/ru/vyarus/guice/persist/orient/support/finder/ObjectDao.groovy
@@ -1,5 +1,7 @@
 package ru.vyarus.guice.persist.orient.support.finder
 
+import com.google.inject.ProvidedBy
+import com.google.inject.internal.DynamicClassProvider
 import com.google.inject.persist.Transactional
 import ru.vyarus.guice.persist.orient.support.finder.mixin.crud.ObjectCrudMixin
 import ru.vyarus.guice.persist.orient.support.finder.mixin.pagination.PaginationMixin
@@ -10,6 +12,7 @@ import ru.vyarus.guice.persist.orient.support.model.Model
  * @since 26.10.2014
  */
 @Transactional
+@ProvidedBy(DynamicClassProvider)
 interface ObjectDao extends ObjectCrudMixin<Model>,
         CustomMixin<Model, String>,
         PaginationMixin<Model, Model> {