+++ src/test/groovy/ru/vyarus/guice/persist/orient/support/finder/inheritance/PowerFinder.groovy
@@ -1,5 +1,7 @@
 package ru.vyarus.guice.persist.orient.support.finder.inheritance
 
+import com.google.inject.ProvidedBy
+import com.google.inject.internal.DynamicClassProvider
 import com.google.inject.persist.Transactional
 import com.orientechnologies.orient.core.record.impl.ODocument
 import ru.vyarus.guice.persist.orient.support.model.Model
@@ -12,6 +14,7 @@ import ru.vyarus.guice.persist.orient.support.model.Model
  * @since 16.10.2014
  */
 @Transactional
+@ProvidedBy(DynamicClassProvider)
 public interface PowerFinder extends BaseFinder1<Model>, BaseFinder2<Model, ODocument>,
         ComplexGeneric<Model, List<Model>>, ComplexGeneric2<Model, Model[], List<Model>> {