+++ src/test/groovy/ru/vyarus/guice/persist/orient/support/modules/FinderTestModule.groovy
@@ -1,17 +1,17 @@
 package ru.vyarus.guice.persist.orient.support.modules
 
 import com.google.inject.AbstractModule
+import ru.vyarus.guice.persist.orient.FinderModule
 import ru.vyarus.guice.persist.orient.support.AutoScanFinderModule
 
 /**
  * @author Vyacheslav Rusakov 
- * @since 04.08.2014
+ * @since 15.12.2014
  */
-@Deprecated
-class AutoScanFinderTestModule extends AbstractModule {
+class FinderTestModule extends AbstractModule {
     @Override
     protected void configure() {
         install(new PackageSchemeModule())
-        install(new AutoScanFinderModule("ru.vyarus.guice.persist.orient.support"));
+        install(new FinderModule());
     }
 }